Refactor mul to use shared methods where possible.
