+++ tensorflow/lite/micro/kernels/cmsis_nn/mul.cc
@@ -14,6 +14,7 @@ limitations under the License.
 ==============================================================================*/
 
 #include "tensorflow/lite/kernels/internal/reference/mul.h"
+#include "tensorflow/lite/micro/kernels/mul.h"
 
 #include "CMSIS/NN/Include/arm_nnfunctions.h"
 #include "tensorflow/lite/kernels/internal/quantization_util.h"
@@ -27,64 +28,7 @@ limitations under the License.
 namespace tflite {
 namespace {
 
-constexpr int kInput1Tensor = 0;
-constexpr int kInput2Tensor = 1;
-constexpr int kOutputTensor = 0;
-
-struct OpData {
-  int32_t output_activation_min;
-  int32_t output_activation_max;
-
-  int32_t output_multiplier;
-  int output_shift;
-
-  // Cached tensor zero point values for quantized operations.
-  int32_t input1_zero_point;
-  int32_t input2_zero_point;
-  int32_t output_zero_point;
-
-  float output_activation_min_f32;
-  float output_activation_max_f32;
-};
-
-TfLiteStatus CalculateOpData(TfLiteContext* context, TfLiteNode* node,
-                             TfLiteMulParams* params, OpData* data) {
-  const TfLiteTensor* input1 = GetInput(context, node, kInput1Tensor);
-  TF_LITE_ENSURE(context, input1 != nullptr);
-  const TfLiteTensor* input2 = GetInput(context, node, kInput2Tensor);
-  TF_LITE_ENSURE(context, input2 != nullptr);
-  TfLiteTensor* output = GetOutput(context, node, kOutputTensor);
-  TF_LITE_ENSURE(context, output != nullptr);
-
-  TF_LITE_ENSURE_EQ(context, NumInputs(node), 2);
-  TF_LITE_ENSURE_EQ(context, NumOutputs(node), 1);
-
-  TF_LITE_ENSURE_TYPES_EQ(context, input1->type, input2->type);
-
-  if (output->type == kTfLiteInt8) {
-    TF_LITE_ENSURE_STATUS(CalculateActivationRangeQuantized(
-        context, params->activation, output, &data->output_activation_min,
-        &data->output_activation_max));
-
-    double real_multiplier = static_cast<double>(input1->params.scale) *
-                             static_cast<double>(input2->params.scale) /
-                             static_cast<double>(output->params.scale);
-    QuantizeMultiplier(real_multiplier, &data->output_multiplier,
-                       &data->output_shift);
-
-    data->input1_zero_point = input1->params.zero_point;
-    data->input2_zero_point = input2->params.zero_point;
-    data->output_zero_point = output->params.zero_point;
-  } else {
-    CalculateActivationRange(params->activation,
-                             &data->output_activation_min_f32,
-                             &data->output_activation_max_f32);
-  }
-
-  return kTfLiteOk;
-}
-
-void EvalQuantized(TfLiteContext* context, TfLiteNode* node, const OpData* data,
+void EvalQuantized(TfLiteContext* context, TfLiteNode* node, const OpDataMul* data,
                    const TfLiteEvalTensor* input1,
                    const TfLiteEvalTensor* input2, TfLiteEvalTensor* output) {
   tflite::ArithmeticParams op_params = {};
@@ -101,7 +45,6 @@ void EvalQuantized(TfLiteContext* context, TfLiteNode* node, const OpData* data,
       tflite::micro::GetTensorShape(input1),
       tflite::micro::GetTensorShape(input2), &op_params);
 
-  if (input1->type == kTfLiteInt8) {
     if (need_broadcast) {
       reference_integer_ops::BroadcastMul4DSlow(
           op_params, tflite::micro::GetTensorShape(input1),
@@ -122,95 +65,33 @@ void EvalQuantized(TfLiteContext* context, TfLiteNode* node, const OpData* data,
                                tflite::micro::GetTensorShape(input2),
                                tflite::micro::GetTensorShape(output)));
     }
-  } else if (input1->type == kTfLiteInt32) {
-    if (need_broadcast) {
-      reference_ops::BroadcastMul4DSlow(
-          op_params, tflite::micro::GetTensorShape(input1),
-          tflite::micro::GetTensorData<int32_t>(input1),
-          tflite::micro::GetTensorShape(input2),
-          tflite::micro::GetTensorData<int32_t>(input2),
-          tflite::micro::GetTensorShape(output),
-          tflite::micro::GetTensorData<int32_t>(output));
-    } else {
-      reference_ops::Mul(op_params,
-                                 tflite::micro::GetTensorShape(input1),
-                                 tflite::micro::GetTensorData<int32_t>(input1),
-                                 tflite::micro::GetTensorShape(input2),
-                                 tflite::micro::GetTensorData<int32_t>(input2),
-                                 tflite::micro::GetTensorShape(output),
-                                 tflite::micro::GetTensorData<int32_t>(output));
-    }
-  }
-}
-
-void EvalFloat(TfLiteContext* context, TfLiteNode* node,
-               TfLiteMulParams* params, const OpData* data,
-               const TfLiteEvalTensor* input1, const TfLiteEvalTensor* input2,
-               TfLiteEvalTensor* output) {
-  tflite::ArithmeticParams op_params = {};
-  op_params.float_activation_min = data->output_activation_min_f32;
-  op_params.float_activation_max = data->output_activation_max_f32;
-
-  bool need_broadcast = reference_ops::ProcessBroadcastShapes(
-      tflite::micro::GetTensorShape(input1),
-      tflite::micro::GetTensorShape(input2), &op_params);
-
-  if (need_broadcast) {
-    reference_ops::BroadcastMul4DSlow(
-        op_params, tflite::micro::GetTensorShape(input1),
-        tflite::micro::GetTensorData<float>(input1),
-        tflite::micro::GetTensorShape(input2),
-        tflite::micro::GetTensorData<float>(input2),
-        tflite::micro::GetTensorShape(output),
-        tflite::micro::GetTensorData<float>(output));
-  } else {
-    reference_ops::Mul(op_params, tflite::micro::GetTensorShape(input1),
-                       tflite::micro::GetTensorData<float>(input1),
-                       tflite::micro::GetTensorShape(input2),
-                       tflite::micro::GetTensorData<float>(input2),
-                       tflite::micro::GetTensorShape(output),
-                       tflite::micro::GetTensorData<float>(output));
-  }
 }
 
 }  // namespace
 
-void* InitMul(TfLiteContext* context, const char* buffer, size_t length) {
-  TFLITE_DCHECK(context->AllocatePersistentBuffer != nullptr);
-  return context->AllocatePersistentBuffer(context, sizeof(OpData));
-}
-
-TfLiteStatus PrepareMul(TfLiteContext* context, TfLiteNode* node) {
+TfLiteStatus Eval(TfLiteContext* context, TfLiteNode* node) {
   TFLITE_DCHECK(node->builtin_data != nullptr);
   auto* params = reinterpret_cast<TfLiteMulParams*>(node->builtin_data);
 
   TFLITE_DCHECK(node->user_data != nullptr);
-  OpData* data = static_cast<OpData*>(node->user_data);
-
-  return CalculateOpData(context, node, params, data);
-}
-
-TfLiteStatus EvalMul(TfLiteContext* context, TfLiteNode* node) {
-  TFLITE_DCHECK(node->builtin_data != nullptr);
-  auto* params = reinterpret_cast<TfLiteMulParams*>(node->builtin_data);
-
-  TFLITE_DCHECK(node->user_data != nullptr);
-  const OpData* data = static_cast<const OpData*>(node->user_data);
+  const OpDataMul* data = static_cast<const OpDataMul*>(node->user_data);
 
   const TfLiteEvalTensor* input1 =
-      tflite::micro::GetEvalInput(context, node, kInput1Tensor);
+      tflite::micro::GetEvalInput(context, node, kMulInput1Tensor);
   const TfLiteEvalTensor* input2 =
-      tflite::micro::GetEvalInput(context, node, kInput2Tensor);
+      tflite::micro::GetEvalInput(context, node, kMulInput2Tensor);
   TfLiteEvalTensor* output =
-      tflite::micro::GetEvalOutput(context, node, kOutputTensor);
+      tflite::micro::GetEvalOutput(context, node, kMulOutputTensor);
 
   switch (input1->type) {
     case kTfLiteInt8:
-    case kTfLiteInt32:
       EvalQuantized(context, node, data, input1, input2, output);
       break;
+    case kTfLiteInt32:
+      EvalQuantizedReference(context, node, data, input1, input2, output);
+      break;
     case kTfLiteFloat32:
-      EvalFloat(context, node, params, data, input1, input2, output);
+      EvalFloatReference(context, node, params, data, input1, input2, output);
       break;
     default:
       TF_LITE_KERNEL_LOG(context, "Type %s (%d) not supported.",
@@ -222,10 +103,10 @@ TfLiteStatus EvalMul(TfLiteContext* context, TfLiteNode* node) {
 }
 
 TfLiteRegistration Register_MUL() {
-  return {/* Init=*/InitMul,
+  return {/* Init=*/MulInit,
           /* Free=*/nullptr,
-          /* Prepare=*/PrepareMul,
-          /*invoke=*/EvalMul,
+          /* Prepare=*/MulPrepare,
+          /*invoke=*/Eval,
           /*profiling_string=*/nullptr,
           /*builtin_code=*/0,
           /*custom_name=*/nullptr,
